wip, subscription streams required
